wav2lip_api/requirements_old.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
anyascii==0.3.3
anyio==4.9.0
asgiref==3.9.0
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
babel==2.17.0
backoff==2.2.1
bangla==0.0.5
bcrypt==4.3.0
beautifulsoup4==4.13.5
billiard==4.2.1
blinker==1.9.0
blis==1.2.1
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
build==1.3.0
cachetools==5.5.2
catalogue==2.0.10
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
chromadb==1.0.20
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpathlib==0.21.1
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.2.1
coqpit==0.0.17
cycler==0.12.1
cymem==2.0.11
Cython==3.1.2
dateparser==1.1.8
decorator==4.4.2
distro==1.9.0
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
docopt==0.6.2
durationpy==0.10
einops==0.8.1
encodec==0.1.1
exceptiongroup==1.3.0
ffmpeg-python==0.2.0
filelock==3.19.1
Flask==3.1.1
flatbuffers==25.2.10
fonttools==4.58.5
frozenlist==1.7.0
fsspec==2025.5.1
future==1.0.0
g2pkk==0.1.2
gdown==5.2.0
google-auth==2.40.3
googleapis-common-protos==1.70.0
grpcio==1.73.1
gruut==2.2.3
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
gTTS==2.5.4
h11==0.16.0
hangul-romanize==0.1.0
hf-xet==1.1.5
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
inflect==7.5.0
itsdangerous==2.2.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.6
joblib==1.5.1
jsonlines==1.2.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
kombu==5.5.4
kubernetes==33.1.0
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
librosa==0.8.1
llvmlite==0.43.0
marisa-trie==1.2.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.8.4
mdurl==0.1.2
mmh3==5.2.0
more-itertools==10.7.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
murmurhash==1.0.13
networkx==2.8.8
nltk==3.9.1
num2words==0.5.14
numba==0.60.0
numpy==1.22.0
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
oauthlib==3.3.1
ollama==0.5.1
onnxruntime==1.22.1
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
orjson==3.11.3
overrides==7.7.0
packaging==25.0
pandas==1.5.3
pillow==11.3.0
platformdirs==4.3.8
pooch==1.8.2
posthog==5.4.0
preshed==3.0.10
proglog==0.1.12
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
PyJWT==2.10.1
pynndescent==0.5.13
pyparsing==3.2.3
PyPika==0.48.9
pypinyin==0.54.0
pyproject_hooks==1.2.0
pysbd==0.3.4
PySocks==1.7.1
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.5
requests-oauthlib==2.0.0
resampy==0.4.3
rich==14.0.0
rpds-py==0.27.1
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.11.4
sentence-transformers==2.2.2
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.7
soxr==0.5.0.post1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.3
srsly==2.5.1
SudachiDict-core==20250515
SudachiPy==0.6.10
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
thinc==8.3.4
threadpoolctl==3.6.0
tifffile==2025.5.10
tokenizers==0.13.3
tomli==2.2.1
torch==2.8.0
torchaudio==2.7.1
torchvision==0.23.0
tqdm==4.67.1
trainer==0.0.36
transformers==4.33.1
triton==3.4.0
TTS==0.22.0
typeguard==4.4.4
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
umap-learn==0.5.7
Unidecode==1.4.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
vine==5.1.0
wasabi==1.1.3
watchfiles==1.1.0
wcwidth==0.2.13
weasel==0.4.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.20.1
zipp==3.23.0
